requirements.txt
python-dotenv
grpcio
grpcio-tools
avro
requests